modified:   home/admin.py 	modified:   home/apps.py 	modified:   home/migrations/0022_account_salt.py 	modified:   home/migrations/0023_auto_20210220_1706.py 	modified:   home/migrations/0024_token.py 	modified:   home/migrations/0025_auto_20210319_1101.py 	modified:   home/models.py 	modified:   home/serializers.py 	modified:   home/tests.py 	modified:   home/urls.py 	modified:   home/views.py 	modified:   product/migrations/0001_initial.py 	modified:   product/migrations/0002_remove_category_keywords.py 	modified:   product/migrations/0003_auto_20210110_1158.py 	modified:   product/migrations/0004_auto_20210110_1223.py 	modified:   product/migrations/0005_auto_20210110_1410.py 	modified:   product/migrations/0006_remove_comment_status.py 	modified:   product/migrations/0007_auto_20210112_0105.py 	modified:   product/migrations/0008_auto_20210112_0107.py 	modified:   product/migrations/0009_delete_comment.py 	modified:   product/migrations/0010_comment.py 	modified:   product/migrations/0011_slider.py 	modified:   product/migrations/0012_auto_20210203_1045.py 	modified:   product/migrations/0013_delete_comment.py 	modified:   product/migrations/0014_comment.py 	modified:   product/migrations/0015_productfromshopee.py 	modified:   product/migrations/0016_auto_20210307_0005.py
	product/serializers.py
	static/js/vendor/bootstrap.bundle.min.js.map
	staticfiles/

+++ product/migrations/0005_auto_20210110_1410.py
@@ -13,11 +13,11 @@ class Migration(migrations.Migration):
         migrations.AlterField(
             model_name='product',
             name='price',
-            field=models.FloatField(),
+            field=models.DecimalField(max_digits=12, decimal_places=2),
         ),
         migrations.AlterField(
             model_name='variants',
             name='price',
-            field=models.FloatField(),
+            field=models.DecimalField(max_digits=12, decimal_places=2),
         ),
     ]
